+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.18.3 (Aug 15, 2025)
-* Added `request.ExtractScheme` to retrieve the request scheme, taking into account headers.
-* Added `request.ExtractHost` to retrieve request hostname, taking into account headers.
+* Added `host` to structured logging attributes.
+* Improved `scheme` in structured logging attributes to parse `Forwarded` HTTP header.
 
 # 0.18.2 (May 13, 2025)
 * Added support for `jsonapi` `links` and `meta` for list responses.